* fix the dll export issue when building the plugin dlls
First, the extern "C" should be added, otherwise, the
wxDYNLIB_FUNCTION(CreatePluginFunctionType, CreatePlugin, *dll);
can't find the function named by "CreatePlugin" in the dlls

Second, the define WXEXPORT __attribute__((dllexport)) should be defined
Since the wx 3.0 header file of dlimpexp.h does not define the WXEXPORT
correctly, we have to copy a new WXEXPORT definition from the git header
of the dlimpexp.h, which define the WXEXPORT correctly.

Third, the __WXMSW__ is enough for the Windows OS, so simplify the #if
condition.